requirements.txt
# Flask and Extensions
Flask>=2.0.0
flask-socketio>=5.0.0
Werkzeug>=2.0.0
Jinja2>=3.0.0
click>=8.0.0
python-engineio
python-socketio
# WebSocket and Async
eventlet>=0.30.0
python-engineio>=4.0.0
python-socketio>=5.0.0

flask-limiter>=3.5.0
flask-socketio>=5.3.6
python-dotenv>=1.0.0
# SSH and System
paramiko>=2.7.0
psutil

# Utilities
requests>=2.26.0
cryptography>=3.4.0
python-dateutil>=2.8.2
netifaces>=0.11.0